Vista Propietario Departamento

+++ Condominios/Condominios/src/main/java/com/proyecto/repository/PropietarioRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.proyecto.entity.Propietario;
 import com.proyecto.entity.Usuario;
 
-public interface PropietarioRepository extends JpaRepository<Propietario, Integer>{
+public interface PropietarioRepository extends JpaRepository<Propietario, Long>{
 
 }
